Crate nom_utils

Structs

ParseError

TagParser

Struct to support displaying better errors for tag()

Enums

ParseErrorKind

Functions

explain_nom

make_context_error

make_span

tag

Like nom::bytes::complete::tag, except that we print what the tag was expecting if there was an error. I feel like this should be the default behavior TBH.

Type Aliases

IResult

Span